rename Card to NavigationCard, create Card
