+++ src/app/login/components/form-login/form-login.module.css
@@ -1,9 +1,3 @@
-.formLayout {
-  border-radius: 6px;
-  padding: 2rem;
-  box-shadow: 0 0 20px 10px #f1eeee;
-}
-
 .loginForm {
   min-width: 360px;
   margin: 1rem;
