+++ src/components/navigation-card/navigation-card.module.css
